Fix some compiler warnings.
git-svn-id: https://svn.isais.edf.fr/mfee/saturne/trunk@9051 e44c5d27-f830-0410-8ce8-974f7a764be6

+++ src/base/cs_c_bindings.f90
@@ -876,7 +876,7 @@ module cs_c_bindings
       bind(C, name='cs_gui_physical_model_select')
       use, intrinsic :: iso_c_binding
       implicit none
-      integer, intent(out) :: ieos, ieqco2
+      integer(c_int), intent(out) :: ieos, ieqco2
     end subroutine cs_gui_physical_model_select
 
     !---------------------------------------------------------------------------
